refactor(web): entities/auth を entities/session へリネーム
認証再編の語彙に合わせ、FSD の entity スライス名も更新する。セッション状態
(useSession) を中心に、共有の turnstile・バリデーション・認証エラー文言を束ねる
エンティティとして session と命名する。
