+++ application/apps/web/src/client/entities/session/model/notify-fetch-error.ts
