+++ application/apps/web/src/client/entities/session/model/validation.ts
