+++ application/apps/web/src/client/features/article/hooks/use-articles.ts
@@ -13,7 +13,7 @@ import { useLocation, useSearchParams } from 'react-router'
 import { toast } from 'sonner'
 import useSWR from 'swr'
 import { useIsMobile } from '@/client/components/shadcn/hooks/use-mobile'
-import { dismissFetchError, notifyFetchError, TOAST_ID } from '@/client/entities/auth'
+import { dismissFetchError, notifyFetchError, TOAST_ID } from '@/client/entities/session'
 import createSWRFetcher from '@/client/infrastructure/create-swr-fetcher'
 import { ALL_MEDIA, isAllMediaSelected, type SelectedMedia } from '../model/media-selection'
