+++ application/apps/web/src/client/features/sessions/error-message.test.ts
@@ -1,4 +1,4 @@
-import { AUTH_ERROR_MESSAGES } from '@/client/entities/auth'
+import { AUTH_ERROR_MESSAGES } from '@/client/entities/session'
 import { resolveLoginErrorMessage } from './error-message'
 
 describe('resolveLoginErrorMessage', () => {
