+++ application/apps/web/src/client/infrastructure/create-swr-fetcher.ts
@@ -1,6 +1,6 @@
 import { fetchWithTimeout } from '@trend-diary/runtime/http'
 import { ClientError, ServerError } from '@trend-diary/std/errors'
-import { notifySessionExpired } from '@/client/entities/auth'
+import { notifySessionExpired } from '@/client/entities/session'
 import getApiClientForClient from '@/client/infrastructure/api'
 
 interface ApiCallResponse {
